refactor: useFav
