+++ src/hooks/useFav.ts
@@ -1,6 +1,7 @@
 import React, { useState } from 'react';
 
 import { reParseSearch } from '@APM/utils/re';
+import { useNoxSetting } from '@APM/stores/useApp';
 import usePlaylist from './usePlaylist';
 
 /**
@@ -10,7 +11,9 @@ import usePlaylist from './usePlaylist';
  */
 const useFav = (playlist: NoxMedia.Playlist) => {
   const [rows, setRows] = useState<NoxMedia.Song[]>([]);
+  const updatePlaylist = useNoxSetting((state) => state.updatePlaylist);
   const { updateSubscribeFavList } = usePlaylist();
+  const saveCurrentList = () => updatePlaylist(playlist);
 
   const handleSearch = (searchedVal: string) => {
     if (searchedVal === '') {
@@ -31,6 +34,7 @@ const useFav = (playlist: NoxMedia.Playlist) => {
     setRows,
     handleSearch,
     rssUpdate,
+    saveCurrentList,
   };
 };
